spacecraft wip stable v2

+++ app/src/main/java/com/ardovic/farkle/dice/game/Coordinate.kt
@@ -1,3 +1,3 @@
 package com.ardovic.farkle.dice.game
 
-class Coordinate(var x: Int, var y: Int)
+class Coordinate(var x: Float, var y: Float)
